feat(message) : breakout async process runner from previously refactored process runner
This will allow moving the async version over time
